fix: 컨설팅 답글 작성하기 스텝1 수정
