+++ src/features/posts/types/consulting-response-form-values.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../constants/consulting-response-form-field-name';
-import { FACE_TYPE } from '../ui/consulting-response-form/consulting-response-form-step-2';
-import { HAIR_TYPE } from '../ui/consulting-response-form/consulting-response-form-step-3';
+import { FACE_TYPE } from '../ui/consulting-response-form/consulting-response-form-step-1';
+import { HAIR_TYPE } from '../ui/consulting-response-form/consulting-response-form-step-2';
 
 const FACE_TYPE_VALUES = [
   FACE_TYPE.OVAL,
@@ -20,13 +20,12 @@ const HAIR_TYPE_VALUES = [
 ] as const;
 
 export const consultingResponseFormSchema = z.object({
-  [CONSULTING_RESPONSE_FORM_FIELD_NAME.option1]: z.string().optional(),
-  [CONSULTING_RESPONSE_FORM_FIELD_NAME.option2]: z.enum(FACE_TYPE_VALUES),
-  [CONSULTING_RESPONSE_FORM_FIELD_NAME.option3]: z.enum(HAIR_TYPE_VALUES).nullable().optional(),
+  [CONSULTING_RESPONSE_FORM_FIELD_NAME.option1]: z.enum(FACE_TYPE_VALUES),
+  [CONSULTING_RESPONSE_FORM_FIELD_NAME.option2]: z.enum(HAIR_TYPE_VALUES).nullable().optional(),
+  [CONSULTING_RESPONSE_FORM_FIELD_NAME.option3]: z.string().optional(),
   [CONSULTING_RESPONSE_FORM_FIELD_NAME.option4]: z.string().optional(),
   [CONSULTING_RESPONSE_FORM_FIELD_NAME.option5]: z.string().optional(),
   [CONSULTING_RESPONSE_FORM_FIELD_NAME.option6]: z.string().optional(),
-  [CONSULTING_RESPONSE_FORM_FIELD_NAME.option7]: z.string().optional(),
 });
 
 export type ConsultingResponseFormValues = z.infer<typeof consultingResponseFormSchema>;
