+++ src/features/posts/ui/consulting-response-form/consulting-response-form.tsx
@@ -1,32 +1,24 @@
 import type { FormStep } from '@/shared/type/form-step';
+import type { KeyOf } from '@/shared/type/types';
+import MultiStepForm from '@/shared/ui/multi-step-form';
+import { FormProvider, useForm } from 'react-hook-form';
 import { CONSULTING_RESPONSE_FORM_FIELD_NAME } from '../../constants/consulting-response-form-field-name';
 import type { ConsultingResponseFormValues } from '../../types/consulting-response-form-values';
-import { FormProvider, useForm } from 'react-hook-form';
-import MultiStepForm from '@/shared/ui/multi-step-form';
-import type { KeyOf } from '@/shared/type/types';
 import ConsultingResponseFormStep1 from './consulting-response-form-step-1';
 import ConsultingResponseFormStep2 from './consulting-response-form-step-2';
-import ConsultingResponseFormStep3 from './consulting-response-form-step-3';
 
 const CONSULTING_RESPONSE_FORM_STEPS: FormStep<ConsultingResponseFormValues>[] = [
   {
     name: CONSULTING_RESPONSE_FORM_FIELD_NAME.option1,
-    question: '컨설팅 제목을 입력하세요',
-    description: '작성하지 않으면 고민 유형에 따라 자동으로 제목이 부여됩니다',
-    required: false,
+    question: '고객님의 얼굴형을 골라주세요',
+    required: true,
     children: <ConsultingResponseFormStep1 />,
   },
   {
     name: CONSULTING_RESPONSE_FORM_FIELD_NAME.option2,
-    question: '고객님의 얼굴형을 골라주세요',
-    required: true,
-    children: <ConsultingResponseFormStep2 />,
-  },
-  {
-    name: CONSULTING_RESPONSE_FORM_FIELD_NAME.option3,
     question: '고객님의 모발 타입을 골라주세요',
     required: false,
-    children: <ConsultingResponseFormStep3 />,
+    children: <ConsultingResponseFormStep2 />,
   },
 ];
 
@@ -36,6 +28,10 @@ export default function ConsultingResponseForm() {
   console.log('method.getValues()', method.getValues());
 
   const canMoveNext = (name: KeyOf<ConsultingResponseFormValues>) => {
+    if (name === CONSULTING_RESPONSE_FORM_FIELD_NAME.option1) {
+      const value = method.getValues(name);
+      return !!value;
+    }
     return true;
   };
